+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+1.0.2 / 2020-07-11
+------------------
+
+- Added `pica` options support (`alpha`, `unsharpAmount`, `unsharpRadius`,
+  `unsharpThreshold`, `cancelToken`).
+
+
 1.0.1 / 2020-06-25
 ------------------